fix: SEAL leaf u64 consistency + digest integrity tests
- compute_feedback_leaf_v1 now uses u64 for feedback_index (consistent with revoke/response leaves)
- Remove u32 truncation cast in give_feedback
- Add digest integrity tests: correct seal_hash verification, forged seal_hash detection, forged client detection
- Update docs: remove archived validation module references, update architecture diagram

+++ programs/agent-registry-8004/src/reputation/instructions.rs
@@ -171,7 +171,7 @@ pub fn give_feedback(
     let leaf = compute_feedback_leaf_v1(
         &asset_bytes,
         &client_bytes,
-        feedback_index as u32,
+        feedback_index,
         &seal_hash,
         slot,
     );
